+++ examples/angular/src/app/shared/components/edit-contentlet-button/edit-contentlet-button.component.ts
@@ -1,8 +1,6 @@
 import { Component, HostBinding, input } from '@angular/core';
 
-import { DatePipe, NgOptimizedImage } from '@angular/common';
-
-import { Contentlet } from '../../models';
+import { Contentlet } from '../../contentlet.model';
 import { DotCMSShowWhenDirective } from '@dotcms/angular/next';
 import { editContentlet } from '@dotcms/uve';
 import { UVE_MODE } from '@dotcms/types';
